index.html added
This is my Tic Tac Toe game to generate and using HTML, CSS and JAVASCRIPT. it will help me to improve our skill and strong command in JavaScript.
The simplicity of Tic-Tac-Toe makes it a popular and easy-to-learn game for all ages. Despite its straightforward nature, the game can involve strategic thinking, especially when playing against a skilled opponent.